Fixing bug #82294 (moon image reversed in southern hemisphere).  The problem was actually not limited to the moon; all object images were flipped left/right in the southern hemisphere.
I traced the problem to three mistakes in the code that computes the "position
angle" of an image; the rotation angle for drawing the image on screen.  These
three mistakes canceled each other out for northern hemisphere locations, but
combined to produce the reported bug for southern hemisphere locations.

The position angle code was duplicated in two places: one for planet images,
and one for deep-sky symbols and images.  Both instances of this code were
wrong in the same way, so deep-sky object images were similarly flipped in the
southern hemisphere.  The code is no longer duplicated; all objects now call
SkyMap::findPA() to get their position angle.

CCMAIL: 82294-done@bugs.kde.org
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=318280

+++ kstars/deepskyobject.cpp
@@ -223,7 +223,7 @@ void DeepSkyObject::drawSymbol( QPainter &psky, int x, int y, int PositionAngle,
 	}
 }
 
-void DeepSkyObject::drawImage( QPainter &psky, int x, int y, int PositionAngle, double zoom, double scale ) {
+void DeepSkyObject::drawImage( QPainter &psky, int x, int y, double PositionAngle, double zoom, double scale ) {
 	QImage *image = readImage();
 	QImage ScaledImage;
 	
@@ -237,7 +237,7 @@ void DeepSkyObject::drawImage( QPainter &psky, int x, int y, int PositionAngle, 
 			ScaledImage = image->smoothScale( w, h );
 			psky.save();
 			psky.translate( x, y );
-			psky.rotate( double( PositionAngle ) );  //rotate the coordinate system
+			psky.rotate( PositionAngle );  //rotate the coordinate system
 			psky.drawImage( -dx, -dy, ScaledImage );
 			psky.restore();
 		}
